Refactor folders mostly on navigation components
