+++ src/components/navigations/Community/Drawer/CommunityMenu.svelte
@@ -10,7 +10,7 @@
 
 <script lang="ts">
   import SelectIconButton from './SelectIconButton.svelte';
-  import type {CommunityType} from '../../../types';
+  import type {CommunityType} from '../../../../types';
 
   export let community: CommunityType;
   export let isSelected = false;
