+++ src/components/navigations/Community/Drawer/CommunityPlusMenu.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
   import SelectIconButton from './SelectIconButton.svelte';
-  import {SvgPlus} from '../../../utils/Icon';
+  import {SvgPlus} from '../../../../utils/Icon';
 
   export let redirectPath: string;
   export let onSelectAddCommunity: (id: string) => void;
