+++ src/components/navigations/Community/Drawer/index.svelte
@@ -54,8 +54,8 @@
 <script lang="ts">
   import {onMount} from 'svelte';
   import {_} from 'svelte-i18n';
-  import type {CommunityType} from '../../../types';
-  import {SvgChevronsLeft, SvgMenu} from '../../../utils/Icon';
+  import type {CommunityType} from '../../../../types';
+  import {SvgChevronsLeft, SvgMenu} from '../../../../utils/Icon';
   import CommunityMenu from './CommunityMenu.svelte';
   import CommunityPlusMenu from './CommunityPlusMenu.svelte';
   import Menu from './Menu.svelte';
