+++ src/components/navigations/Community/index.svelte
@@ -23,12 +23,12 @@
 </style>
 
 <script lang="ts">
-  import Dashboard from '../pages/dashboard/Index.svelte';
+  import Dashboard from '../../pages/dashboard/Index.svelte';
   import Router from 'svelte-spa-router';
-  import Drawer from '../uis/drawer/Index.svelte';
-  import Temp from '../pages/Temp.svelte';
-  import Member from '../pages/Member/Member.svelte';
-  import Feed from '../pages/Feed/Feed.svelte';
+  import Drawer from './Drawer/index.svelte';
+  import Temp from '../../pages/Temp.svelte';
+  import Member from '../../pages/Member/Member.svelte';
+  import Feed from '../../pages/Feed/Feed.svelte';
 
   const communites = [
     {
