+++ src/components/navigations/Main/Header/Manipulation.svelte
@@ -4,6 +4,7 @@
     align-items: center;
     justify-content: end;
   }
+
   .search,
   .notification {
     margin-right: 25px;
@@ -14,23 +15,40 @@
       display: none;
     }
   }
+
   .profile {
     cursor: pointer;
   }
 </style>
 
 <script lang="ts">
+  import type {definitions} from '../../../../types/supabase';
+
   import {push, replace} from 'svelte-spa-router';
-  import supabase from '../../../lib/db';
-  import {user} from '../../../stores/sessionStore';
-  import {SvgBell} from '../../../utils/Icon';
+  import {user} from '../../../../stores/sessionStore';
+  import {SvgBell} from '../../../../utils/Icon';
   import Profile from './Profile.svelte';
   import Search from './Serach.svelte';
+  import {onMount} from 'svelte';
+  import supabase from '../../../../lib/db';
 
-  const signInOut = async (): Promise<void> => {
+  onMount(async () => {
     if ($user) {
-      await supabase.auth.signOut();
-      await replace('/');
+      let {data, error, status} = await supabase
+        .from<definitions['User']>('User')
+        .select(`displayName, name, avatar_url`)
+        .eq('id', $user.id)
+        .single();
+
+      console.log('data', data);
+
+      if (error && status !== 406) throw error;
+    }
+  });
+
+  const goToProfile = async (): Promise<void> => {
+    if ($user) {
+      await replace('/profile');
 
       return;
     }
@@ -46,7 +64,7 @@
   <div class="notification">
     <SvgBell />
   </div>
-  <div class="profile" on:click={signInOut}>
+  <div class="profile" on:click={goToProfile}>
     <Profile name="hanna" imageSrc={'https://picsum.photos/200'} />
   </div>
 </div>
