+++ src/components/navigations/Main/index.svelte
@@ -5,21 +5,22 @@
     grid-template-rows: 56px 1fr;
   }
   @media (--mobile) {
-    .container {}
+    .container {
+    }
   }
 </style>
 
 <script>
   import Router from 'svelte-spa-router';
-  import Profile from '../pages/Profile.svelte';
-  import Header from '../uis/header/index.svelte';
-  import SignIn from '../pages/SignIn.svelte';
-  import SignUp from '../pages/SignUp.svelte';
-  import ServiceSample from '../pages/ServiceSample.svelte';
-  import Onboarding from '../pages/Onboarding.svelte';
-  import CommunityCrete from '../pages/community/Create.svelte';
-  import CommunityRoute from './Community.svelte';
-  import Temp from '../pages/Temp.svelte';
+  import Profile from '../../pages/Profile.svelte';
+  import Header from './Header/index.svelte';
+  import SignIn from '../../pages/SignIn.svelte';
+  import SignUp from '../../pages/SignUp.svelte';
+  import ServiceSample from '../../pages/ServiceSample.svelte';
+  import Onboarding from '../../pages/Onboarding.svelte';
+  import CommunityCreate from '../../pages/community/CommunityCreate.svelte';
+  import CommunityRoute from '../Community/index.svelte';
+  import Temp from '../../pages/Temp.svelte';
 
   const routes = {
     '/sign_in': SignIn,
@@ -28,7 +29,7 @@
     // '/lorem/:repeat': Lorem,
     '/profile': Profile,
     '/service_sample': ServiceSample,
-    '/community/create': CommunityCrete,
+    '/community/create': CommunityCreate,
     '/community': CommunityRoute,
     '/community/*': CommunityRoute,
     '/*': Temp,
